requirements.txt
# Core
Django==4.2.10
python-dotenv==1.0.0
python-decouple==3.8
dj-database-url==2.0.0
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.9

# API
djangorestframework==3.14.0
django-filter==23.1
djangorestframework-simplejwt==5.3.1
# djangorestframework-api-key==2.2.0

# CORS
django-cors-headers==4.7.0

# HTTP
urllib3<2.0.0

# File handling
Pillow==10.2.0

# Gunicorn
gunicorn==21.2.0

# Development
debugpy==1.8.0
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
django-stubs==4.2.1

# Required for deployment
setuptools>=68.0.0
wheel>=0.40.0
